Split up header functions/classes into separate file
